sistema para crear usuario

+++ src/main/java/com/intesoft/syncworks/domain/entities/Rol.java
@@ -1,12 +1,12 @@
 package com.intesoft.syncworks.domain.entities;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table(name = "rol")
 public class Rol {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
 }
